streamlit_app/requirements.txt
streamlit>=1.35.0
ultralytics>=8.2.0
torchvision>=0.18.0
opencv-python-headless>=4.9.0.80
Pillow>=10.3.0
requests>=2.31.0
numpy>=2.0.0
scikit-learn>=1.6.0
pandas>=2.2.2
